fix: refactor api handlers, update docs
