+++ Cargo.toml
@@ -5,7 +5,7 @@ resolver = "2"
 [workspace.package]
 authors      = ["arrow-services <group-services@arrowair.com>"]
 edition      = "2021"
-homepage     = "https://www.arrowair.com/docs/documentation/services/intro"
+homepage     = "https://www.aetheric.nl"
 license-file = "LICENSE"
 repository   = "https://github.com/aetheric-oss/svc-contact"
