+++ client-grpc/Cargo.toml
@@ -37,7 +37,7 @@ tower       = { version = "0.4", optional = true }
 [dependencies.lib-common]
 features = ["grpc"]
 git      = "https://github.com/aetheric-oss/lib-common.git"
-tag      = "latest-develop"
+tag      = "main"
 
 [dependencies.tokio]
 features = ["macros", "rt-multi-thread"]
