+++ client-rest/Cargo.toml
@@ -26,7 +26,7 @@ tokio = { version = "1.33", features = ["full"] }
 [dev-dependencies.lib-common]
 features = ["grpc"]
 git      = "https://github.com/aetheric-oss/lib-common.git"
-tag      = "latest-develop"
+tag      = "main"
 
 [[example]]
 name = "rest"
